multi_model_chat_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi_model_chat_agent"
version = "0.1.0"
description = "A multi-model chat agent that queries GPT-4o, GPT-4o-mini, and Claude in parallel for cost dashboard testing"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk>=0.5.3",
    "scale-gp",
    "openai",
    "opentelemetry-api",
    "opentelemetry-sdk",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
